+++ IsoSpec++/cwrapper.h
@@ -116,7 +116,7 @@ void deleteFixedEnvelope(void* tabulator, bool releaseEverything);
 const double* massesFixedEnvelope(void* tabulator);
 const double* probsFixedEnvelope(void* tabulator);
 const int*    confsFixedEnvelope(void* tabulator);
-int confs_noFixedEnvelope(void* tabulator);
+size_t confs_noFixedEnvelope(void* tabulator);
 
 double empiricAverageMass(void* tabulator);
 double empiricVariance(void* tabulator);
